+++ x/pocketcore/types/merkle_test.go
@@ -43,7 +43,7 @@ func TestEvidence_GenerateMerkleRoot(t *testing.T) {
 		},
 		TotalRelays: 5,
 		Proofs: []Proof{
-			{
+			RelayProof{
 				Entropy:            3238283,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -51,7 +51,7 @@ func TestEvidence_GenerateMerkleRoot(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            34939492,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -59,7 +59,7 @@ func TestEvidence_GenerateMerkleRoot(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            12383,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -67,7 +67,7 @@ func TestEvidence_GenerateMerkleRoot(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            96384,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -75,7 +75,7 @@ func TestEvidence_GenerateMerkleRoot(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            96384812,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -128,7 +128,7 @@ func TestEvidence_GenerateMerkleProof(t *testing.T) {
 		},
 		TotalRelays: 5,
 		Proofs: []Proof{
-			{
+			RelayProof{
 				Entropy:            3238283,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -136,7 +136,7 @@ func TestEvidence_GenerateMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            34939492,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -144,7 +144,7 @@ func TestEvidence_GenerateMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            12383,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -152,7 +152,7 @@ func TestEvidence_GenerateMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            96384,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -160,7 +160,7 @@ func TestEvidence_GenerateMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            96384812,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -217,7 +217,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 		},
 		TotalRelays: 5,
 		Proofs: []Proof{
-			{
+			RelayProof{
 				Entropy:            83,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -225,7 +225,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            3492332332249492,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -233,7 +233,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            121212123232323383,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -241,7 +241,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            23121223232396384,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -249,7 +249,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            963223233238481322,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -267,7 +267,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 		},
 		TotalRelays: 9,
 		Proofs: []Proof{
-			{
+			RelayProof{
 				Entropy:            82398289423,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -275,7 +275,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            34932332249492,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -283,7 +283,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            1212121232383,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -291,7 +291,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            23192932384,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -299,7 +299,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            2993223481322,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -307,7 +307,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            993223423981322,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -315,7 +315,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            90333981322,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -323,7 +323,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            2398123322,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),
@@ -331,7 +331,7 @@ func TestEvidence_VerifyMerkleProof(t *testing.T) {
 				Token:              validAAT,
 				Signature:          "",
 			},
-			{
+			RelayProof{
 				Entropy:            99322342381322,
 				SessionBlockHeight: 1,
 				ServicerPubKey:     nodePubKey.RawString(),