+++ doc/changelog.md
@@ -6,3 +6,4 @@
 - Changed RPC and from `node-proof` to `node-receipt`
 - Update posmint module to use sdk.Ctx interface
 - Fix `pseudorandomGenerator` unexported properties would return empty json
+- Evidence now holds proof interface to allow for challenge proofs